Massive Update--Read Description
Finished getMovieInfo() function in JS file; fixed all searching
problems and the problem we had with not getting the right movie.
Everything works well with that; all that’s left is to push the info we
have onto the HTML page, and I created a function for each bit of info
for us to do that. Also created a skeleton for the YouTube API function.